backend/requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
langchain
langchain-google-genai
python-dotenv
psycopg2-binary
json_repair
sqlalchemy
